requirements.txt
Babel==2.7.0
Django==2.2.5
django-phonenumber-field==3.0.1
phonenumbers==8.10.19
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
